WATER-1131: Handling gauging station 404
Handles gauging stations that are linked to from a license that don't
currently have any data. So where the open data return a 404.

Instead of showing an error page, the gauging station page is still
shown, but with a message telling the user that there is no data.

 - Changes the view-licences helper to return a null object if the
station returns a 404.
 - Adds a unit test for this.
 - Changes the importing of the connector, to import the full object to
allow sinon to stub the required functions.