Bug-fix: Collect all validation errors
Changes the behaviour of Joi to collect all the validation errors rather
than existing on the first error.

This means that the forms will show all the errors at once to the user.

In this change, this effects the notification templates and the two
forms that collect the user's contact information prior to sending a
notification.