Fix: Shows email validation error
Use the `formError` if the email address is not valid.

Shortcuts the controller handler to prevent processing of bad input and
displays the feedback to the user.

Changes `getUser` to deal with numeric ids. No longer need to lowercase
the value either because this happens in the pre query filter in the
IDM.

Gets the user via the email address and application name to handle
multiple IDM entries for the same email address.

Adds form validation plugin validators which are required so the
formError is not truthy, covering page, sort and direction.