Fix: Unauthorised agent access to return
Adds route permissions to prevent an agent accessing a return that they
are not authorised to view.

Filter document roles when getting licence numbers to only primary_user
and user_returns roles. This is in addition to the route permissions to
prevent a user with access to multiple companies accessing unauthorised
licence returns.