Fix: Corrects meter reading validation
WATER-1502

Fixes a bug with the validation of meter readings by changing the
validation schema to be created based on the values that are submitted.

The schema function now has to receive the data in order to create the
rules to enforce that a non null value is:

1) Greater than the meter start reading
2) Greater than the most recent non null reading

Also corrects sign in tests. This would have been overlooked because the
previous meter readings check in left an `only` clause in the tests so
only that test was executed.