Feat: Implement additional security headers
WATER-1554

Implements CSP in report only mode.
Adds feature-policy header.
Migrates inline CSS to SCSS files

WATER-1420

Removes the JS that disabled the button on the change access page to
align better with GDS. This script was due to have a nonce added to
comply with CSP, so it could just be removed instead.