Fix: Includes lic holder name in security letter
WATER-1699

Fixes an issue with the licence holder name not appearing the letter
that contains a security code.

There are 6 available slots for name and address. Plus a separate slot
of postcode.

There are up to 8 address lines in a licence, so if all 8 are present,
the AddressLine4 data is omitted.