Fix: Cookie too large for user with many roles
WATER-1730

When an agent becomes responsible for many licences, they in turn have
more items in the entity_roles array found when signing in.

This data is all serialized into the cookie, and once the user hits
three/four role pairs (user, user_returns) the amount of data being
added to the cookie for a given domain is too large. (Including GA as
well).

Therefore this quick fix maps the entity_roles picking out only the
necessary data. E.g. for 4 roles the length of the session data went
from ~4000 to ~1000 bytes.

There is a technical improvement story coming soon to only store a
session id in the cookie. But this approach is fast to implement and
will hopefully unlock users sooner.