Feat: Use accessible autocomplete for long selects
When a select element contains more than 200 items and the user permits
client side JS, the select is replaced with an accessible autocomplete
component.

Adds the govuk-body class to the form-group that contains the select to
allow the style to cascade into the list of options that replace the
select options.