Feat: Dynamically load internal user scopes
WATER-1691

Extension of previous work for external users, this time dealing with
the internal users, whose permissions are based on scopes that used to
be included in the cookie.

This meant that the sign out - sign in cycle had to be completed to see
updated roles take effect.

Introduces a new plugin that operates in the onCredentials phase where
the user is loaded and the `request.auth.credentials` object is updated
with the user's scopes.

Also updates the handlebars pagination helper to make is handle an
undefined pagination object.

Adds licence loader plugin

Allows some licence data to be added to the request before the handler
is executed. This allows common licence data to be requested to save
duplicate code in the controller functions.

Makes use of the above plugin to update the mainNav functions that
determine which links to place in the top nav bar. Now the options are
reduced for a user who has no registered licences.