Fix: Prevent redirect of users to welcome page
WATER-1955

Fixes a bug where the user could end up on a page that is not the
company selector without having selected a company. This mean licence
lookups where not authorised so the user was redirected to the welcome
page.

This change updates the `auth-credentials` plugin to add more content to
the request under the `defra` property (to avoid collisions). This data
is then used in a `preResponse` handler to redirect the user either to
the `select-company` route if the user has companies, or the
`add-licences` route if they don't.

Also hides the company switcher if there is no company name selected
preventing it showing if the user has not selected a company but wishes
to update a password.