Allow internal returns user to edit returns with 'completed' status (#641)
Feat: Allow internal returns user to edit returns with 'completed' status

WATER-1954

Change logic to allow internal returns users to edit returns which are not void, are after summer 2018 and are past the end date. This is to include returns with a 'completed' status. Fix test to match logic change