Feat: Changes to paper form issuing
WATER-1950

Allow paper forms to be produced even if they have already been received
by updating the query filter to also included 'complete' returns and not
to enforce that the return isCurrent flag is true.

Add licence ended reasons for notification preview to warn the user that
the licence is revoked/expired/lapsed.