Fix: Attempts to fix an auth error
Some users are being shown the `application unavailable` page via Nginx
and this fixes an issue that would arise if the session store and the
cookie session ids were out of sync.

In the event that the session id does not resolve to a session in the
database, the cookie is cleared and the user is sent back to the welcome
page.