Removes macro for notify status badge
Replaces with badge-status function that maps notify status to a known
badge status and text that is then rendered by the badge macro.

Adds tests for the sort-query and sort-new-direction filters.

Adds tests for the controller to ensure that the messages are augmented
with the badge status data ready for render.