Fix: Update to fix of rest api
A recent addition to the hapi rest api means that the error logging
fails if the required method signature is not satisfied.

This change updates to a later fixed version of the hapi rest api.

I don't have the permission to publish to npm so this change links
directly to the git version which can be changed later.