Feat: Change email UI
WATER-2154

Update email UI and company selector update

 - Includes new account-setting landing page
   - Update email link commented out at this stage
 - Includes confirm password page
 - Adds page for adding new email address
 - Adds page for email verification code
 - Adds success page
 - Adds the try again page for failed password attempts

Reworks the logic for the company selector which can now use the
`request.defra` object to help determine if a user should be redirected,
and also adds a settings item to allow a route to opt out of company
selector redirection. This came about because the user is logged in when
updating their account details so the existing mechanism of checking if
the route did not require authentication was no longer relevant.