Feat: Adds user journey info to error reporting
WATER-2166

Adds a shared plugin that records the request made by a user to session
to help with bug fixing when errors are logged.

Changes the logger setup so that the logger is monkey patched with a
method for logging errors with user journeys.

Uses the new logger method in all places where the HAPI Request object
is available.

Updates dependencies to handle some security advisories from npm.

Some tests have been updated where there were now issues using
`server.inject` because the requests were attempting to make calls to
the water service.