Fix: Fixes password verification step of update password
WATER-2236

Fixes an issue whereby the password verification step could be satisfied
by supplying a password that meets the password format criteria, but did
not have to actually be the user's password.

Whilst fixing that the html forms have moved to nunjucks from
handlebars.

Also, the URLs for the update-password flow have been changed to sit
under `/account/update-password` and the routes include 301 redirects
from the existing URLs.