Feat: Refine/reduce error logs
WATER-2212

Reduces the noise in the errbit logs.

1) Downgrade a incorrect email as part of the reset password from error
to debug.

2) Downgrade an add licence error to a info log (LicenceSimilarityError,

LicenceMissingError)
3) Handles cleared session in the add licence flow

Deals with the user pressing back or refresh at the end of the add
licence flow which would attempt to extract data from a null session.
Instead now the user is redirected back to the start of the flow.

4) Adds more logging for map condition error

Current logging does not supply enough information to diagnose this so
added more logging to the `mapCondition` function to aid a fix in the
future.