Fix: Fixes back link behaviour in company selector
WATER-2202

Updates the company selector so that the back link is only rendered if
the user has already selected a company. In which case they can be sent
back to the licences page.

Also updates @hapi/hapi to 18.4.0 for security reasons which in turn require    d an
update to @hapi/hoek