Fix: Allow _ga query param on some routes
WATER-2420

Fixes an issue where an authenticated user coming from the site via
https://www.gov.uk/guidance/manage-your-water-abstraction-or-impoundment-licences-online
gets a 400 response due to the unexpected `_ga` query param.

Fixes a sorting error in the licences page where each of licence
name, licence number and end date, all actually sorted by licence
number.

Adds a bit more logging to the error logger so that more of the request
details are available when diagnosing an issue.