Feat: Adds skeleton markup for billing summary page
WATER-2435

Creates the markup for the billing summary page using a test object.

Adds the batchService local to the billing module which can be used for
service layer composition of data ready for the controller to use.

Also adds the cancel/confirm batch screens but there are also not
completely functional yet and just redirect to the batch list at the
moment.