Feat: Integrate backend batch APIs
WATER-2447

Updates the UI to get batch information from the water service instead
of a stubbed response.

The water service is still implementing the mechanism to get the value
and number of invoices in a batch so for now, for the purpose of viewing
the design, the numbers are randomised.

Then updates the batch summary to use real invoice data from the water
service.