Notes on current mob position
We started the day by trying to follow what `src/internal/modules/billing/controllers/create-bill-run.js::postBillingBatchRegion()` actually does.

This meant digging into how the UI is generating a request to the `water-abstraction-service` using code in the `water-abstraction-helpers`. This was all because we followed the rabbit hole of `const { regions } = request.pre;` (whole method beloew for context).

```javascript
const postBillingBatchRegion = async (request, h, refDate) => {
  const { regions } = request.pre;
  const schema = billingRegionFormSchema(regions);
  const billingRegionForm = forms.handleRequest(selectBillingRegionForm(request, regions), request, schema);

  if (!billingRegionForm.isValid) {
    const { selectedBillingType, selectedTwoPartTariffSeason } = forms.getValues(billingRegionForm);
    const path = getRegionUrl(selectedBillingType, selectedTwoPartTariffSeason);
    return h.postRedirectGet(billingRegionForm, path);
  }

  try {
    const batch = getBatchDetails(request, billingRegionForm, refDate);
    const { data } = await services.water.billingBatches.createBillingBatch(batch);
    const path = routing.getBillingBatchRoute(data.batch, { isBackEnabled: false });
    console.log("I am not ready to be watched!!")
    return h.redirect(path);
  } catch (err) {
    if (err.statusCode === 409) {
      return h.redirect(getBatchCreationErrorRedirectPath(err));
    }
    throw err;
  }
};
```

We learnt that where you see a reference to `request.pre` we should find a pre handler on the route definition and that is how, for example, `request.pre.regions` gets populated.

We also learnt in `water-abstraction-helpers` we're wrapping the request to an API in an `EventEmitter`. It looks like this was added to allow the UI to tag the user ID onto all http requests by listening for a `ON_PRE_REQUEST` event, hooking in and amending the request body with the ID.

The second line `const schema = billingRegionFormSchema(regions);` seems to be building a Joi schema to ready to be used to validate the data coming in on the request to the UI.

We finished on the next line; `const billingRegionForm = forms.handleRequest(selectBillingRegionForm(request, regions), request, schema);`. Another rabbit hole, we spotted that the project is using [Hapi vision](https://hapi.dev/module/vision/) to decorate the server, request and response. It does this in a plugin (don't have the reference to hand anymore!)

At that point we called it a day 😁