Notes on current mob position Today we continued our dive into `src/internal/modules/billing/controllers/create-bill-run.js::postBillingBatchRegion()` and finished our look into the line `const billingRegionForm = forms.handleRequest(selectBillingRegionForm(request, regions), request, schema);`.
We followed through `selectBillingRegionForm` which uses `formFactory` to build the form to be displayed. We learnt that fields can have a mapper, which translates data in the form into data we expect, and vice-versa -- for example, turning the date returned by a date widget into the date format we want. The different fields available to us (which live in `src/shared/lib/forms/fields/`) have a default mapper, which can be overridden when the field is added to a form -- for example, mapping a yes/no set of radio buttons to true/false using `booleanMapper`.

Once we'd figured out what `selectBillingRegionForm` was doing, we looked into `forms.handleRequest()`, which does the job of validating the data in the form.

It starts by reading the validation type from the form and loading the appropriate validation adapter. This will be either joi or json-schema. Since the form we're looking at uses Joi, we decided to leave json-schema for another day!

Once we have the validation adapter, we clone the form (so we can make changes to it without affecting the original) and pull in the form data from the payload.

We then determine the schema to validate against -- if a schema was passed in to `handleRequest()` then we use that, otherwise we use the validation adapter's `createSchemaFromForm()` to build a schema based on the form's fields (eg. a field that uses `booleanMapper` will be validated against `Joi.boolean()`).

We validate the data in the request against the schema. We use `getCustomErrors()` to pull in any custom errors defined on the form -- in this case, the `selectedBillingRegion` radio field has a custom error of `Select the region` if Joi returns the `any.required` error.

The validation adapter's `formatErrors()` method is used to take any errors returned by Joi and reformat them into a friendlier format. It also applies any custom errors that were defined, eg. our `Select the region` error.

We apply these errors to the cloned form and set the form's `isValid` property based on whether any errors were returned.

Next, we encounter `postValidate()`. This basically takes each field, calls the `postValidate` hook of the field's mapper (passing in the current value of the field), and sets the field's value to whatever was returned. For the majority of mappers, `postValidate` simply returns the data as-is using the lodash `identity` function (which caused some head-scratching as to its purpose!). The only time this does anything different is for dates, where the value is reformatted. This is all needed because a date validated by Joi becomes a Date object, and therefore needs to be changed back to the expected format.

Finally, the field values in the cloned form are set to whatever we got back from `postValidate()`, and this amended form is returned by `handleRequest()`.

The work done yesterady and today means we now know exactly what the first 3 lines of `postBillingBatchRegion()` do! And we can move onto the fourth line in our next mob session.