Connect to new financial years endpoint
We're not quite ready to re-write how things are done so we've added a method to the existing `BillingBatchService` which handles connecting to our new endpoint in the water-abstraction-service. This follows the existing pattern

Note - Just realised all the service files (plus I can see others) use `TitleCase` in the files names!! I've always worked on the convention all file names are always lower and include no spaces to make working with them in the terminal easier (using snake or kebab case depending on community). But at the very least be consistent across your own code base! 😩