Use our new fin. years endpoint to drive journey
Now we have our endpoint we can use its return value to determine if we should direct to the page where we select the financial year to bill or not.

We're also on the cusp of using the returned data to generate the options on the UI.

All still quick & dirty though hence plethora of 'naughty-naming'!