Update getBillingBatchFinancialYear use std. view
One of the things that _really_ takes some getting your head round on this project is it's custom form builder. Add to that we have found ocasions where a 'formbuilder-lite' pattern is used (see use of `formRender` in a number of views which leads to `src/shared/view/nunjucks/macros/forms/index.njk`) and the standard GDS design system.

@cruikshanks has determined that if we are to pick one, it should be the standard GDS design system pattern because

- those of us who have used Nunjucks before did it in this way
- it makes it much easier to correllate what we see in the views with the examples GDS provide
- in the future, we should be able to build a delivery pipeline where the designers prototype can form the basis of the final version using a simple copy and paste from the prototype kit

Certainly, whilst the GDS design system is full documented and there are plenty of examples from other services to call on, the formBuilder has no documentation and only its inconsistent usage in this project to go on.

This change is our first steps into amending the `getBillingBatchFinancialYear()` controller to return a standard view rather than one generated using the formBuiler.

We have been using `src/internal/views/nunjucks/notifications/step.njk` heavily as a reference for how to do this within Water.