Reorganise create-bill-run controller file
One of the issues we have had when dealing with this file is that functions and controllers are mixed together. It is also not clear what is a controller and what is a function.

This change trys to rectify the problems by moving all the functions to the bottom and ensuring the `get` and `post` controllers are grouped together. To further make the distinction clear we have started to use the convention of `_` prefixed to functions/methods that are intended to be private to the module. We also dropped any `get` prefixes as they were conflicting with the `get` controllers.