Move remaining controllers up + renames
Further to the previous clean up we have moved some other controllers we found up into the controller section, and what we thought was a controller into the functions. We also renamed the functions so that the naming allowed related functions to be grouped together which will hopeful help folks in the future.