Add billable years page in 2-Part tarriff bill run journey (#2104)
https://eaflood.atlassian.net/browse/WATER-3679

By default when creating a 2-part tariff bill run the current financial year is built. The users have identified scenarios where previous financial years have not been billed. This means they currently cannot generate the bill runs for them.

The plan is to do a check within the create bill run journey for available billable financial years and if more than one is found, take the user to a screen where they select the one they wish to build.

If only one is found the journey will remain unaffected and the user will never be shown the page.

We're reliant on the existing functionality to determine the current financial year for a 2-part tariff bill run (batch). This gets sent to the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service/pull/1695) which uses it to determine what `unsent` billable years there are. This result is

- what decides whether the page gets displayed
- the options to display in the new page

Some other things we have determined whilst working on this ticket

- all new changes will use what we are referring to as the GDS standard way of creating views i.e. we will avoid using the form builder and the form builder macros
- functions not exported from a module should be prefixed with `_`. This is a common convention to denote private methods and it makes it clearer when reading the code how a function is to be used.
- in controller modules controllers should be at the top with functions below
- controller `post` and `get` functions should be grouped together

And finally, people who stub the Hapi `response` object should be made to sit and think about what they have done 😁 .