Remove `no-undef` rule overrides
The `no-undef` rule stops errors being raised if an undefined variable is used. This was being used in some of the Cypress test files to avoid errors on the Cypress globals such as `cy`. We have added these to the `globals` section of the standard config in `package.json` so we don't need these overrides any more.