Remove `no-useless-escape` overrides
The rule `no-useless-escape` flags up places where an escape character is being used unnecessarily. There are a few regexes in the code which have these unneeded escape chars, with the rule being disabled for these lines. (Note that in some cases, the rule was being disabled for a line which didn't actually contain an unnecessary escape!). We can't see any good reason to keep these uneeded escapes, so we remove the overrides and fix the escapes.