Refactor tests for _getInvoiceAccountData()
Were moving all the tests of the private methods into one that tests the actual output of `createCSV()`. There is a prinicple of one assertion per test. However, where we are just mass testing the properties of an object under a single scenario we believe that adds unnecessary noise to the test file. So, we start grouping our assertions into a single new test.