Move to set values in expectations
Noted in the last couple of changes that we were specifying the values in the expectations rather than referencing our test objects. This is because the logic is peforming various formatting on some of the values. To replicate this we would have to call the same underlying formatters.

We felt it was cleaner in these cases to be explicit with the expeced values. Then we decided if we're doing it there we may as well be consistent and use explicit values throughout.