Add test for min charge dropped columns
If you read the logic when the transaction has a charge type of minimum charge we actually drop a whole host of columns. This is a very disconcerting as we're supposed to be generating a CSV file (the columns should still exost but just be empty). We should be making this clear in the unit tests. So, we add a test to cover this specific scenario and ascertion.

We drop everything in the existing test. We think they were attempting to assert the same scenario but possibly weren't aware that by checking the columns they expect to still exist they weren't confirming the ones that shouldn't were gone.