Deal with a 'leaky' test
The test for 'Adjusted source type' being tidal was found to be failing. Upon investigation we realised it was a problem we have encountered before.

For example, invoice was declared at the start of testing and then as the tests run various changes are being made to it and its children. You might assume that the object is reset for each test bit that is not the case. Changes made for a test near the top of the file will be seen by tests that come after. Such was the case with this, the experiment for minimum charge was amending the transaction charge type and causing the logic in `transactions-csv.js` to skip adding the column our tidal test relied on.

We solved this problem in [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api/blob/main/test/support/helpers/general.helper.js) by ensuring we cloned any object we then intened to amend for a test. This stopped changes leaking between tests.

To do this we created a helper and placed it in `test/support/`. Our convention on that project was anything used solely to support testing, such as fixtures and helper functions should be placed there. It's a convention we intend to bring to WRLS.