Remove the use of 'clever' reduce
If you are approaching JavaScript as a functional language or you are doing stuff client side using reduce _might_ make sense. As we're on the server side though, the performance differences are irrelevant and the code clearly shows it has been written in an imperative style.

So, it comes down to readability and `reduce()` fails utterly on that. Reduce is intended to be used to generate an aggregate value, for example, sum of a hundred numbers. How its been used here is just 'clever' for its own sake. It avoids declaring a separate const and instantiates the empty array with _slightly_ less code. But for anyone who isn't totally clued up on JavaScript idiosyncrasies, such as those new to the language or juniors, it's an unreadable mess!