Remove _getTransactionAmounts()
We have had to add 3 new functions in order to stick with our aim of creating a function that makes it clear what columns will be exported rather than calling other functions to return them, and using the spread operator.

The craziness of having separate debit and credit fields rather than single value and type fields forces us to do clever checks around whether something is a credit or a debit. On top of that someone decided for the transaction lines to return a message rather than `null` hence the need for a special method to handle those fields.

Hopefully though everything is still readable.