Start removing getTransactionData()
Some git funnies mean our changes for starting to remove the getTransactiondata function have been merged with the change to update the chargeReason method to a function (rather than a const).

Suffice to say, we have come up with a way of handling minimum charge transactions which cause a number of columns to be dropped without resorting to calling a separate method and then spreading the result.