Tidy up rowToStrings()
This was a method split across 3 methods for 'reasons', one of which is no longer required. It just didn't make sense and is certainly not readable.

We've ditched what was there before and migrated it to a standard, 'single' function. Ideally we'd have ditched lodash altogether but our heads hurt and this is all we can manage today.