Update back to where we left it
We started making some changes, refactoring the code as part of implementing the SROC version of the CSV file. That was until we realised we needed a wholesale tidy up of this module.

We did that in https://github.com/DEFRA/water-abstraction-ui/pull/2120 but rebasing this branch was going to be easier if we dropped the changes we made.

This commit puts back those not now covered by the tidy up.