Fix export bill run csv for sroc (#2143)
* Fix `volume`

After discussion we have agreed that the `volume` field in the output csv should be the transaction's `volume` value

* Fix `Abatement factor`

The `Abatement factor` field wasn't being populated in the output CSV. On investigation we realised that this is because we don't get it back from the Charging Module! However there is a reason for this. The CM passes on the value it gets from the Rules Service; and the Rules Service only returns values that it itself has calculated. For other factors, where we supply a boolean, it therefore does calculation and returns a value. For the abatement factor, we provide it with the value to use, and therefore it doesn't do any calculation of a factor.

All this means that in this instance (unlike other factors such as S127, we use the factor that we were initially given rather than the one we received back (which is usually held in the appropriate `calc` field)

Co-authored-by: Stuart Adair <43574728+StuAA78@users.noreply.github.com>