Fix field for sroc
The alcs export relies on the transaction's  `isCompensationCharge` boolean to determine if this is a compensation charge or not. We copied this across to sroc. However in testing we found that this is not actually present, possibly because the mapper which sets this field is being bypassed. We therefore use its underlying logic (check whether `chargeType` is `compensation`) to determine if this line is a compensation charge.