Fix Compensation Charge field in CSV export (#2145)
* Fix Compensation Charge field in CSV export

https://eaflood.atlassian.net/browse/WATER-3586

The Compensation Charge field is not correctly populated with `Y` for compensation charges. This is due to the expected `isCompensationCharge` flag not being present. We therefore replicate the logic of this flag in our CSV exporter.

* Fix field for sroc

The alcs export relies on the transaction's  `isCompensationCharge` boolean to determine if this is a compensation charge or not. We copied this across to sroc. However in testing we found that this is not actually present, possibly because the mapper which sets this field is being bypassed. We therefore use its underlying logic (check whether `chargeType` is `compensation`) to determine if this line is a compensation charge.