Add Delete All Billing Data environment variable
For integration test purposes we need to be able to use the Delete All Bill Runs function in the `preprod` environment.

The logic was previously "Only enable this for specific values of `NODE_ENV`". We change this to use its own flag, `ENABLE_DELETE_ALL_BILLING_DATA_BUTTON`, and as a failsafe we also check that the environment is not `production`.