Postal address issue - remove unneeded postcode from template https://eaflood.atlassian.net/browse/WATER-3717
When reviewing the UI changes for the postal address issue we realised that although we had added `personalisation.postcode` to the template, this wouldn't actually be supplied by the service -- instead the postcode would come back as one of the numbered address lines. We therefore remove this from the template.