Addition of mBOD as a unit type in the tagging flow
https://eaflood.atlassian.net/browse/WATER-3616

There are licences that have hands-off levels measured in mBOD. mBOD = metres below ordinance datum. As an Environment Officer, I want to tag licences with hands-off levels measured in metres below ordinance datum, so that I can send water abstraction alerts to these licences.

This adds mBOD to the unit of measurement drop down when specifying a threshold as part of the tagging flow.