SRoC charge creation - additional characters required in charge reference description (#2159)
https://eaflood.atlassian.net/browse/WATER-3700

As part of the SRoC charge creation process, users need to enter a description for the charge reference. B&D want to be able to use ; (semi colon) and ' (apostrophe) in this description.

The list of allowed characters in WRLS is currently limited to letters (lower and upper case), hyphen (dash), numbers, forward slash, ampersand (&), asterisk , commas, brackets.

The following characters are not accepted by SOP “ ? ^ £ ≥ ≤ — (long dash).

We should allow ; and ' in WRLS, in addition to the currently accepted characters. Or we could amend to a 'not allowed' list instead, and only exclude those characters not accepted by SOP.