Skip public water question page if licence holder isn't water undertaker (#2163)
* Skip "supply of public water" question if not a water undertaker
https://eaflood.atlassian.net/browse/WATER-3702

In the SRoC charge information flow, we want to skip the question "Is abstraction for the supply of public water?" if the customer is not a water undertaker.

* Refactor `postChargeCategoryStep`

We were having difficulty parsing `postChargeCategoryStep`. We've refactored it a little so it makes a bit more sense

* Add redirection

We add code to the get request handler to skip the public water page if this isn't a water undertaker

* Update unit tests

Two unit tests were failing because we needed to amend the initial test setup. We fix this and add an additional unit test to ensure our redirection works.

* Additional refactor

We do another refactor sweep to remove some unneeded complexity and nested `if` statements

* Set chargeElement.isSupplyPublicWater to false if isSupplyPublicWater is undefined

This will be set to undefined if isWaterUndertaker is false

* Omit Change link for if not a water undertaker

If the licence holder is not a water undertaker then we want to omit the Change link for "Supply public water" from the review screen. We do this by passing `isWaterUndertaker` to the `view` template, which in turns passes it to the `charge-element-table-sroc` template, where we amend the existing logic of "Display change link if this page is editable" to be "Display change link if this page is editable and the licence holder is a water undertaker"

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>