The charge ref is incorrect in the transaction detail (#2172)
https://eaflood.atlassian.net/browse/WATER-3733

When viewing the transactions as part of an annual bill run for SRoC, there is a section for each transaction showing charge period, charge reference etc.

The value of the charge reference, shown in brackets after the charge reference is displaying incorrectly. The conversion of the value from pence to pounds looks to be incorrect.