Fix Notify callback exception
When a Notify request is generated, Notify hits a callback endpoint to update the request status. The callback endpoint is on the production server which means that any non-prod requests (ie. a password reset on `pre`) will still result in the callback endpoint being hit. This should be fine, as the request id won't exist in prod and therefore nothing would be changed. Unfortunately, our callback endpoint was written without any error handling for such errors and therefore these requests from Notify would cause a small interruption while the server restarts after the unhandled error!

We therefore add simple error handling to ensure that such errors are caught without bringing down the production server for a few seconds.