Make charge version max upload consistent
We've spotted that the charge version upload endpoint route specifies `5e+7` for `maxBytes`. Quick conversion means this is 50MB. However, the UI page says the max upload is 20MB.

We don't like an inconsistency! So, this updates the info in the route config to match what we are stating on the upload page.